Попытка вынести логику в ХОК для плеера
