+++ src/components/player/VideoPlayer.jsx
@@ -1,61 +1,20 @@
-import React, {PureComponent} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+import {withStatePlayer} from './withStatePlayer.jsx';
 
-class VideoPlayer extends PureComponent {
-  constructor(props) {
-    super(props);
+const VideoPlayer = (props) => {
+  const {poster} = props;
 
-    this.videoRef = React.createRef();
-    this.delayVideoTimer = null;
-  }
-
-  componentDidMount() {
-    const video = this.videoRef.current;
-
-    video.src = this.props.srcVideo;
-    video.muted = true;
-  }
-
-  componentWillUnmount() {
-    const video = this.videoRef.current;
-    video.src = ``;
-  }
-
-  componentDidUpdate() {
-    const video = this.videoRef.current;
-
-    const playVideo = () => {
-      video.play();
-    };
-
-    if (this.props.isPlaying) {
-      this.delayVideoTimer = setTimeout(playVideo, 1000);
-    } else {
-      if (this.delayVideoTimer) {
-        clearTimeout(this.delayVideoTimer);
-        this.delayVideoTimer = null;
-      }
-      video.load();
-    }
-  }
-
-
-  render() {
-    const {poster} = this.props;
-
-    return (<video ref={this.videoRef}
-      poster={ poster }
-      autoPlay={false}
-      width="280"
-      height="175">
-    </video>);
-  }
-}
+  return <video
+    poster={ poster }
+    autoPlay={ false }
+    width="280"
+    height="175">
+  </video>;
+};
 
 VideoPlayer.propTypes = {
   poster: PropTypes.string.isRequired,
-  srcVideo: PropTypes.string.isRequired,
-  isPlaying: PropTypes.bool.isRequired,
 };
 
-export default VideoPlayer;
+export default withStatePlayer(VideoPlayer);
